+++ trunk/jdem846/src/us/wthr/jdem846/render/ModelDimensions2D.java
@@ -14,13 +14,13 @@ public class ModelDimensions2D
 	private int tileSize;
 	private int dataRows;
 	private int dataColumns;
-	private float sizeRatio;
+	private double sizeRatio;
 	private int outputWidth;
 	private int outputHeight;
-	private float xDim;
-	private float yDim;
-	private float numTilesHorizontal;
-	private float numTilesVertical;
+	private double xDim;
+	private double yDim;
+	private double numTilesHorizontal;
+	private double numTilesVertical;
 	private long tileOutputWidth;
 	private long tileOutputHeight;
 	private long tileCount;
@@ -51,30 +51,30 @@ public class ModelDimensions2D
 		modelDimensions.outputHeight = modelOptions.getHeight();
 
 		if (modelDimensions.dataRows > modelDimensions.dataColumns) {
-			modelDimensions.sizeRatio = (float)modelDimensions.dataColumns / (float)modelDimensions.dataRows;
-			modelDimensions.outputWidth = Math.round(((float) modelDimensions.outputHeight) * modelDimensions.sizeRatio);
+			modelDimensions.sizeRatio = (double)modelDimensions.dataColumns / (double)modelDimensions.dataRows;
+			modelDimensions.outputWidth = (int) Math.round(((double) modelDimensions.outputHeight) * modelDimensions.sizeRatio);
 		} else if (modelDimensions.dataColumns > modelDimensions.dataRows) {
-			modelDimensions.sizeRatio = (float)modelDimensions.dataRows / (float)modelDimensions.dataColumns;
-			modelDimensions.outputHeight = Math.round(((float)modelDimensions.outputWidth) * modelDimensions.sizeRatio);
+			modelDimensions.sizeRatio = (double)modelDimensions.dataRows / (double)modelDimensions.dataColumns;
+			modelDimensions.outputHeight = (int) Math.round(((double)modelDimensions.outputWidth) * modelDimensions.sizeRatio);
 		}
 		
 		log.info("Output width/height: " + modelDimensions.outputWidth + "/" + modelDimensions.outputHeight);
 		
-		float xdimRatio = (float)modelDimensions.outputWidth / (float)modelDimensions.dataColumns;
+		double xdimRatio = (double)modelDimensions.outputWidth / (double)modelDimensions.dataColumns;
 		modelDimensions.xDim = dataPackage.getAvgXDim() / xdimRatio;
 		//dataPackage.setAvgXDim(xDim);
 
-		float ydimRatio = (float)modelDimensions.outputHeight / (float)modelDimensions.dataRows;
+		double ydimRatio = (double)modelDimensions.outputHeight / (double)modelDimensions.dataRows;
 		modelDimensions.yDim = dataPackage.getAvgYDim() / ydimRatio;
 		//dataPackage.setAvgYDim(yDim);
 		//log.info("X/Y Dimension (cellsize): " + xDim + "/" + yDim);
 		
 
-		modelDimensions.numTilesHorizontal = ((float)modelDimensions.dataColumns) / ((float)modelDimensions.tileSize);
-		modelDimensions.numTilesVertical = ((float)modelDimensions.dataRows) / ((float)modelDimensions.tileSize);
+		modelDimensions.numTilesHorizontal = ((double)modelDimensions.dataColumns) / ((double)modelDimensions.tileSize);
+		modelDimensions.numTilesVertical = ((double)modelDimensions.dataRows) / ((double)modelDimensions.tileSize);
 		
-		modelDimensions.tileOutputWidth = Math.round(((float)modelDimensions.outputWidth) / modelDimensions.numTilesHorizontal);
-		modelDimensions.tileOutputHeight = Math.round(((float)modelDimensions.outputHeight) / modelDimensions.numTilesVertical);
+		modelDimensions.tileOutputWidth = Math.round(((double)modelDimensions.outputWidth) / modelDimensions.numTilesHorizontal);
+		modelDimensions.tileOutputHeight = Math.round(((double)modelDimensions.outputHeight) / modelDimensions.numTilesVertical);
 		
 		modelDimensions.tileCount = (int) (Math.ceil(((double)modelDimensions.dataRows / (double)modelDimensions.tileSize)) * Math.ceil(((double)modelDimensions.dataColumns / (double)modelDimensions.tileSize)));
 		
@@ -100,7 +100,7 @@ public class ModelDimensions2D
 	}
 
 
-	public float getSizeRatio()
+	public double getSizeRatio()
 	{
 		return sizeRatio;
 	}
@@ -118,25 +118,25 @@ public class ModelDimensions2D
 	}
 
 
-	public float getxDim()
+	public double getxDim()
 	{
 		return xDim;
 	}
 
 
-	public float getyDim()
+	public double getyDim()
 	{
 		return yDim;
 	}
 
 
-	public float getNumTilesHorizontal()
+	public double getNumTilesHorizontal()
 	{
 		return numTilesHorizontal;
 	}
 
 
-	public float getNumTilesVertical()
+	public double getNumTilesVertical()
 	{
 		return numTilesVertical;
 	}
